Minor Bugxfix: Solves #31
The new sorting order ensures that older employees are overwritten by
newer ones. And that still working employees
(ISNULL(`Beschäftigungsende`)) overwrite former employees.

This is neccessary because VK is not a real primary key. There can only
be one person at a time wit the same VK number. But that one number can
be given to other workers at a later point in time.