Adding the actual email functionality into the user_input class. changed:       src/php/classes/class.iCalendar.php The new function build_ics_roster_cancelled() returns an iCalendar file that has the STATUS:CANCELLED for exactly one roster_item object.
changed: src/php/classes/class.roster_item.php
public function to_email_message_string() returns a textual representation of the roster_item suitable for email notifications.
The start and the end has to be removed, in order to allow for aggreagation of mails.

changed: src/php/classes/class.user_input.php
New private static function write_email_about_changed_roster_to_employees() to loop through the old and the new roster to search for changes.

New private static function save_email_about_changed_roster_to_database() to store the messages for the user inside the database in order to send them later.
The sending of the messages might be part of the maintenance class. It might also be directly be printed to the user upon his next login, or both.

New private static function send_email_about_changed_roster_to_employees() to create an actual email and also send it.
Right now this is experimental and all the mails will just be sent to the administrator to see if it works, how much mail will be created and if there are any errors.