Build an independent test system using PHPUnit
bootstrap.php executes the necessary commands from default.php
There are slight differences between those file regarding the fact, that the
tests are executed in the command line interface php.

pdr_test.sql is a minimal example of a possible database.
It is used to run the tests against it.