Add HTTPS to the installation requirements
Some redirects were enclosed in:
if (++$_SESSION['number_of_times_redirected'] < 3) {...}
This has the goal to not have endless redirects.

I fear that we might completely loose some of those redirects.
If an error/warning/notice is shown to the user, then the header()
function will not work anymore after that. And the session variable
might not be set, causing such a warning depending on the exact PHP
version and config.

/src/php/pages/install_page_check_requirements.php now only performs
its tests once and saves the results in variables. The success
variables are aggregated into one boolean.