Add new options to marginal-employment-hours-list
Overtime JavaScript did not work on fresh years. The balance was not
reachable. Now there is a span containing the balance as data value.

There are five more options for the hours list for the individual
employees. Those options contain multiple months.
+const PDR_YEAR_QUARTER_FIRST = 121;
+const PDR_YEAR_QUARTER_SECOND = 122;
+const PDR_YEAR_QUARTER_THIRD = 123;
+const PDR_YEAR_QUARTER_FOURTH = 124;
+const PDR_YEAR_FULL = 1212;
Also the page was copied into the employee section of the menu.
Perhaps it will be renamed in the future.