Prepare some functions for alternating weeks
Some employees prefer to work on alternating weeks e.g.:
(A)-Week 6 hours a day starting at 08:00 in the morning and
(B)-Week 8 hours a day starting at 10:00 in the morning.

There is now a class alternating_week that contains some functions to
allow calculations and interactions with changing weeks.

The class principle_roster was made as an extension to the class roster
by simply moving functions from that class into the new extension.
Both classes still cannot yet be instantiated into objects. But I plan
to change that in some tome from now.

The class principle_roster now has to be adopted to the new options.
Also the pages describing and editing the principle rosters have to be
modified.