Prepare principle-roster-employee for alternating
I removed the plots. They could be between the tables. They could be
below the tables. They could be an a completely different page. But
are they realy good for something?

The style of the div class main-area had to be changed. The fixed width
of 95% made the border not span all the input fields.
overflow: auto created a scroll-bar at the bottom of the roster table.