Add new class roster_approval
The approval states of the rosters are now handled in one place in one
abstract class.

$List_of_branch_objects = branch::get_list_of_branch_objects();
This is still called in the default.php Using the global variable is
deprecated. The static class array has to be used.
I made one first attempt to replace uses of the global variable.

The installer now also checks, if all the hidden directories are secured
against access from the public.