Add sync from roster table to bar plot
Changes in the roster tables are now synced to the roster bar plot.
The feature is still buggy. For example it does not work well together
with the sync back to the table when the bar plot is changed via drag
and drop.

The JavaSCript on the client side now has a new global variable
Roster_array which holds the $Roster information.
Also a list of employee last names is new. I did not want to add the
whole workforce object, because it may contain sensitive data.

The input elements of the roster table contain identifying information
in the form of data-* attributes. This helps the JavaSCript to know
where it is without using heuristics to identify which column is start
or end of duty and break.

The SVG now uses the viewBox to decide, where the plots start. This
makes the numbers easier to understand. Also it is more intuitive.

The class roster_item implements \JsonSerializable to make it possible
to serialize private and protected variables. Those variables are not
secret, they just should not be changed from outside the class without
changing other connected variables also.