principle-roster-employee.php seems to work now
insert_changed_entries_into_database takes the $Roster and evaluates all the $roster_items. It decides, if updates to existing entries are made or if new entries will be created.

The value valid_from defines when an entry starts to be valid. The value valid_from is only a secondary parameter, which marks the existence of another entry with the same scope in another time.