Add class network_of_branch_offices
The class holds all objects of the class branch.

No method in the class is static. All the work has to be done on an instance of the class.
However, the array $List_of_branch_objects is static.

All the class variables of branch have been turned to private. Right now they can be viewed by __get. But I want to build one get method for each variable.