Selenium TestRosterHoursPage with json roster values
RosterHoursPage.java tests for absences. There are cases, where there is a registered absence, but still values in the roster too. Therefore we have to check .//td[last()] instead of .//td[2]

Error in class.absence.php
The array keys range from 0 to 7 for the values from 1 to 8.
The test in line 287 has to search for the value, not for the key!