Advance user-management.php
user-management.php now has two inputs for employee key and for user status.
Administrators can edit these.
The email adress is visible but can not be modified.

in head.php it is now possible to add javascript files specific for distinct pages.
unsaved-changes-prompt.js allows a page to ask the user before leaving the page if there are unsubmitted changes.

user class is adapted to provide necessary functions for user-management.php

TODO: Tests are missing for this page.